backend/requirements-ee.txt
# Earth Engine Python API (NASADEM рельеф в scripts/ee_nasadem_stats.py)
earthengine-api>=1.0.0
python-pptx>=1.0.2
